chore: upgrade conda env to streamline creating post-processing python model (#100)
Because

our API produces inference output in standardised format when a model is labelled with a specific task such as classification (`TASK_CLASSIFICATION`) and detection (`TASK_DETECTION`). In order to make this happen, a user has to follow design guidelines to create the post-processing python model **from scratch**, so the model output can be converted to the standardised format. 

To streamline the post-processing python model creation process, a user can define the model using the [`triton-python-model` package in the triton-conda-env image](https://github.com/instill-ai/triton-python-model).

This commit

- update `triton-conda-env` to `v0.2.1-alpha`
- update the test dummy models to use `triton-python-model` package to define the post-processing model
- update `docs/model.md` about `Prepare your own model to deploy on VDP`
  - add how to create a post-processing Python model to standardise the output format by inheriting the corresponding post-processing task class
- update `yolov4-onnx-cpu.zip` to use `triton-python-model` package to define the post-process model